requirements.txt
pandas>=2.0.0
reportlab>=4.0.0
twilio>=8.0.0
openpyxl>=3.1.0
flask>=2.3.0
gunicorn>=21.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
weasyprint